requirements.txt
azure-cognitiveservices-speech==1.20.0
azure-ai-textanalytics==5.1.0
requests==2.28.1